prepared generic data comm tests
